gospace: Added dictionary-based code mobility.
This is a weak form for code mobility and is similar to RPC. There is a
small difference with net/rpc:
    - This implementation exposes in principle any method or function.
        * Closures are limited by the facts that declarations must be in
          order and functionality must agree between two different peers.
          If used locally, closures work properly.
    - There happens a lot of translation, some of it can be
      optimized. Do not consider this an optimized solution.
    - Made GetAgg() and QueryAgg() operators of the protocol and not
      user-side functions.

This implementation in principle allows function binding of functions and methods
between any language by using a specific URI format:

    func://<language-name>:<language-version>/<namespace-or-package-name>[.object-or-structure].<function-or-method>

This might of course change, but for now, that makes each function
addressable by an URI.

function/: Function registry and helps for custom reflection handling of
functions.
    - Added Applier interface for defining map-based functionality for
      both tuples and templates, but in principle anything else.
    - Added IsFunc() for testing of functions and/or methods.
    - Added Type() helper combined for both functions and methods.
    - Refactoring of registry.go and added String() for output.
    - Source clean up.

action.go: Fixed external package naming.

message.go: Undo seperation of message into several parts.

If it will become relevant later, it will be reintroduced.

operations.go: Added GetAgg and PutAgg as operations.

pointtopoint.go: Introducing a hack to circumvent certain limitations in
tuple space implementations.

A function registry should not be attached to this structure. This is a
limitation of the current architecture of gospace.

template.go:
    - Added Apply() for iterating a template and applying a
function on each field.
    - Changed template constructor to consider functions.

tuple.go:
    - Added Apply() for iterating a template and applying a
function on each field.
    - Changed pattern matching to consider functions.

space.go: Encapsulate function and fixed external package naming.

space_error.go: Fixed external package naming.

{tuplespace, tuplespace_alt}.go:
    - Added encoding and decoding of
functions through funcEncode() and funcDecode().
    - Added function registry to the tuple space.
        * The point is that each tuple space have their own function
          registry. A global registry is possible, but this can be added
          by creating a custom TupleSpace structure.
    - Added handleGetAgg() and improved handleQueryAgg().
    - Source clean up.

There is a lot of refactoring to do here, for now, take it for what it
is.