gospace: Fixed weak code mobility for remote peers.
README.md: Updated the documentation.

registry.go: Introduced a global registry.

This allows for symmetric peers to execute code remotely, giving a weak
form for code mobility, but also to bind functions between languages on
remote machines.

For now, this is a workaround in the registry implementation, and a
global registry should be removed once peers can exchange capabilities.

operations.go: Introduced operation for retrieving size of a tuple
space.

waitingclient.go: Return a copy of a template instead.

space.go:
    - Rewrote Size().
    - Undoing importing whole packages with '.'.

tuplespace.go:
    - putP now locks on the tuple space when copying a tuple and its
    clients.
        * This has performance issues for now, but is done in order to
          avoid deadlocking.
    - Make a copy when matching before patternmatching in findTuple() and
      findAllTuples.
        * For same reasons, to avoid deadlocking.
    - Introducing PutAgg() operation.
    - If aggregation fails, empty tuples are returnted.
    - Introducing handleSize().
    - Undoing importing of whole packages with '.'.
    - Using a global function registry instead of a local one.

When using goroutines on the tuple space, deadlocking can still
occur. This can happen either due to port exhaustion or due to multiple
write to binding variables. For now, we throttle the implementation
incuring a penalty on performance.

tuplespace_alt.go:
    - Undoing importing of whole packages with '.'.
    - Code clean up and variable simplification.
    - Added logging instead of dumping everything to standard output.
        * Logging makes only sense when you need it.
    - Rewrote PutAgg().
    - Added timeout option for establishConnection().
        * This is to make it possible to timeout on future operations.
    - Put(), Get() and Query() now blocks on first connection.
        * Until semantics are established, a blocking operation blocks
          until connection to a peer is established.