gospace: Added in-memory connections for local spaces.
Due to the current architecture, this should be seen as a workaround and
to avoid port exhaustion locally when working with a space.

README.md: Updated the documentation.

*.go: Code clean up and documentation.

tuplespace_alt: Added in-memory connect.

Instead making use of a real connection, create a pipe internally
and send this through the channel to the tuple space.